refactor(admin): put `writeInAdjudication` view in API
I'm now using it in both the real API server as well as the memory backend for VxAdmin.
