+++ libs/api/src/services/admin/views/index.ts
@@ -1,0 +1,1 @@
+export * as writeInAdjudicationTable from './write_in_adjudication_table';
